refactor(main): extract asset-download IPC into lib/ipc/registerAssetDownloadHandlers
Amp-Thread-ID: https://ampcode.com/threads/T-019e1fcf-7754-703e-9441-11cc2816c591
